System.Metrics.Acceptor

Description

This top-level module will be used by the acceptor app (the app that asks EKG metrics from the forwarder app).

Documentation

runEKGAcceptor

Arguments

:: AcceptorConfiguration
Acceptor configuration.

-> Store
The store all received metrics will be stored in.

-> IO ()