System.Metrics.Configuration

Documentation

data AcceptorConfiguration

Acceptor configuration.

Constructors

AcceptorConfiguration

Fields

acceptorTracer :: !(Tracer IO (TraceSendRecv (EKGForward Request Response)))

The tracer that will be used by the acceptor in its network layer. For more info about tracers please read its documentation.

forwarderEndpoint :: !HowToConnect

The endpoint that will be used to listen to the forwarder.

requestFrequency :: !NominalDiffTime

Specifies how often the acceptor will ask the framework for new metrics. It can be specified as seconds or as fraction of second.

whatToRequest :: !Request

Specifies what to request: all existing metrics or particular metrics.

shouldWeStop :: !(TVar Bool)

TVar that can be used as a brake: if an external thread will set it to True, the acceptor will send MsgDone message to the forwarder and their session will be closed.

data ForwarderConfiguration

Forwarder configuration.

Constructors

ForwarderConfiguration

Fields

forwarderTracer :: !(Tracer IO (TraceSendRecv (EKGForward Request Response)))

The tracer that will be used by the forwarder in its network layer.

acceptorEndpoint :: !HowToConnect

The endpoint that will be used to connect to the acceptor.

reConnectFrequency :: !NominalDiffTime

If the connection with the acceptor will fail, the forwarder will attempt to re-establish the connection after this delay. It can be specified as seconds or as fraction of second.

actionOnRequest :: !(Request -> IO ())

Additional action that will be performed every time the forwarder will receive the request from the acceptor.

data HowToConnect

Specifies how to connect to the peer.

Constructors

LocalPipe !FilePath
Local pipe (UNIX or Windows).

RemoteSocket !Host !Port
Remote socket (host and port).

type Host = Text

type Port = Word16