System.Metrics.Protocol.Acceptor

Description

A view of the EKG forwarding/accepting protocol from the point of view of the client.

For execution, a conversion into the typed protocol is provided.

Documentation

data EKGAcceptor req resp m a where

Please note that the acceptor is a server from the networking point of view: the forwarder establishes network connection with the acceptor. But after the connection is established, the acceptor becomes a client from the interaction point of view: it sends a request for new metrics, the forwarder replies to the acceptor.

Constructors

SendMsgReq :: req -> (resp -> m (EKGAcceptor req resp m a)) -> EKGAcceptor req resp m a

SendMsgDone :: m a -> EKGAcceptor req resp m a

ekgAcceptorPeer :: Monad m => EKGAcceptor req resp m a -> Peer (EKGForward req resp) 'AsClient 'StIdle m a

Interpret a particular action sequence into the client side of the EKGForward protocol.