System.Metrics.Protocol.Forwarder

Documentation

data EKGForwarder req resp m a

Please note that the forwarder is a client from the networking point of view: it establishes network connection with the acceptor. But after the connection is established, the forwarder becomes a server from the interaction point of view: the acceptor sends a request for new metrics, the forwarder replies to the acceptor.

Constructors

EKGForwarder

Fields

recvMsgReq :: req -> m (resp, EKGForwarder req resp m a)

The acceptor sent us a request for new metrics.

recvMsgDone :: m a

The acceptor terminated. Here we have a pure return value, but we could have done another action in m if we wanted to.

ekgForwarderPeer :: Monad m => EKGForwarder req resp m a -> Peer (EKGForward req resp) 'AsServer 'StIdle m a

Interpret a particular action sequence into the server side of the EKGForward protocol.