System.Metrics.ReqResp

Documentation

type MetricName = Text

Each EKG metric has a unique name.

data MetricValue

Metric value that will be transmitted from the forwarder to the acceptor. Please note that EKG.Distribution is not supported yet.

Constructors

CounterValue !Int64
Counter value.

GaugeValue !Int64
Gauge value.

LabelValue !Text
Text label.

data Request

The request for the new metrics. The acceptor will send this request to the forwarder.

Constructors

GetAllMetrics
Get all metrics from the forwarder's local store.

GetMetrics !(NonEmpty MetricName)
Get specific metrics only.

newtype Response

The response with the metrics. The forwarder will send it to the acceptor as a reply for the request. Please note that the list of metrics can be empty (for example, if the forwarder's local store is empty).

Constructors

ResponseMetrics [(MetricName, MetricValue)]